requirements.txt
pytest==8.1.1
setuptools==69.2.0
opentelemetry-sdk==1.23.0
azure-monitor-opentelemetry==1.3.0
